# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_analyze_photo_nutrition_photo_analyze_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_analyze_photo_nutrition_photo_analyze_post",
        "type": "object"
      },
...
      "ChatRequest": {
        "properties": {
          "message": {
            "maxLength": 8000,
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "ChatRequest",
        "type": "object"
      },
      "ChatResponse": {
        "properties": {
          "agent_used": {
            "title": "Agent Used",
            "type": "string"
          },
          "reply": {
            "title": "Reply",
            "type": "string"
          }
        },
        "required": [
          "reply",
          "agent_used"
        ],
        "title": "ChatResponse",
        "type": "object"
      },
...
      "ConversationRead": {
        "properties": {
          "agent_used": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Used"
          },
          "content": {
            "title": "Content",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "id",
          "role",
          "content",
          "agent_used",
          "timestamp"
        ],
        "title": "ConversationRead",
        "type": "object"
      },
...
      "FoodCatalogRead": {
        "properties": {
          "calories_kcal": {
            "title": "Calories Kcal",
            "type": "number"
          },
          "carbs_g": {
            "title": "Carbs G",
            "type": "number"
          },
          "category_tr": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Tr"
          },
          "fat_g": {
            "title": "Fat G",
            "type": "number"
          },
          "fiber_g": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fiber G"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name_en": {
            "title": "Name En",
            "type": "string"
          },
          "name_tr": {
            "title": "Name Tr",
            "type": "string"
          },
          "protein_g": {
            "title": "Protein G",
            "type": "number"
          },
          "sodium_mg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sodium Mg"
          },
          "sugar_g": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sugar G"
          }
        },
        "required": [
          "id",
          "name_tr",
          "name_en",
          "category_tr",
          "calories_kcal",
          "protein_g",
          "carbs_g",
          "fat_g",
          "fiber_g",
          "sugar_g",
          "sodium_mg"
        ],
        "title": "FoodCatalogRead",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MealPhotoRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "detected_items_summary": {
            "title": "Detected Items Summary",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "detected_items_summary",
          "created_at"
        ],
        "title": "MealPhotoRead",
        "type": "object"
      },
...
      "PhotoMealAnalysisRead": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PhotoMealItemRead"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "PhotoMealAnalysisRead",
        "type": "object"
      },
      "PhotoMealItemRead": {
        "properties": {
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/FoodCatalogRead"
            },
            "title": "Candidates",
            "type": "array"
          },
          "estimated_grams": {
            "title": "Estimated Grams",
            "type": "number"
          },
          "food_name": {
            "title": "Food Name",
            "type": "string"
          },
          "is_uncertain": {
            "default": false,
            "title": "Is Uncertain",
            "type": "boolean"
          },
          "matched_food": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FoodCatalogRead"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "food_name",
          "estimated_grams",
          "matched_food",
          "candidates"
        ],
        "title": "PhotoMealItemRead",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "PulseCoach API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/chat": {
      "post": {
        "operationId": "chat_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Chat",
        "tags": [
          "chat"
        ]
      }
    },
    "/chat/clear": {
      "post": {
        "description": "\"Sohbeti Sıfırla\" - GERİ ALINABİLİR (veri sunucuda kalır, bkz.\nconversation_service.soft_clear). Ekran ve koçun bağlamı bu andan\nitibaren temiz sayfa görür, kalıcı silme için bkz. DELETE /chat/history.",
        "operationId": "clear_history_chat_clear_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Clear History",
        "tags": [
          "chat"
        ]
      }
    },
    "/chat/history": {
      "delete": {
        "description": "\"Sohbeti Kalıcı Olarak Sil\" - GERİ ALINAMAZ, bkz.\nconversation_service.hard_delete_history.",
        "operationId": "delete_history_chat_history_delete",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete History",
        "tags": [
          "chat"
        ]
      },
      "get": {
        "operationId": "history_chat_history_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ConversationRead"
                  },
                  "title": "Response History Chat History Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "History",
        "tags": [
          "chat"
        ]
      }
    },
    "/chat/stream": {
      "post": {
        "description": "POST /chat ile aynı tur, Server-Sent Events olarak (2026-09-26): araç\ndurumu, taslak parçaları ve sonda kesin yanıt (bkz. orchestrator.ChatStream).\nHer olay `data: {json}` satırı. İstemci bağlantıyı koparsa tur yine tamamlanır\nve sohbet geçmişine kaydedilir - araçların yazdığı veriyle geçmiş tutarlı kalsın.",
        "operationId": "chat_stream_chat_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Chat Stream",
        "tags": [
          "chat"
        ]
      }
    },
...
    "/nutrition/photo-analyze": {
      "post": {
        "operationId": "analyze_photo_nutrition_photo_analyze_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_analyze_photo_nutrition_photo_analyze_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhotoMealAnalysisRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Analyze Photo",
        "tags": [
          "nutrition-photos"
        ]
      }
    },
    "/nutrition/photo-history": {
      "get": {
        "operationId": "photo_history_nutrition_photo_history_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 30,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MealPhotoRead"
                  },
                  "title": "Response Photo History Nutrition Photo History Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Photo History",
        "tags": [
          "nutrition-photos"
        ]
      }
    },
    "/nutrition/photo-history/{photo_id}": {
      "delete": {
        "operationId": "delete_photo_history_entry_nutrition_photo_history__photo_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "photo_id",
            "required": true,
            "schema": {
              "title": "Photo Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Photo History Entry",
        "tags": [
          "nutrition-photos"
        ]
      }
    },
    "/nutrition/photo-history/{photo_id}/image": {
      "get": {
        "operationId": "photo_history_image_nutrition_photo_history__photo_id__image_get",
        "parameters": [
          {
            "in": "path",
            "name": "photo_id",
            "required": true,
            "schema": {
              "title": "Photo Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Photo History Image",
        "tags": [
          "nutrition-photos"
        ]
      }
    },
...
  }
}